Enhancing Build Configuration and Environment Variable Management (#27)
* add .envrc.example

* Updating Makefile to Change Build Output Directory

* Display Personal Access Token When NEH_DEBUG Is Set to 't'

* Refactor: Rename Function to getNehServerEndpoint for Clarity

* Add NEH_SERVER_ENDPOINT_DEVELOPMENT to .envrc.example

* Update Environment Variable Check for Localhost Configuration

- Modify the condition to check if the WORKING_ON_LOCALHOST environment variable is set to "t" instead of any non-empty value.
- This ensures that the application correctly identifies when it is running in a localhost environment.

* Implement Environment Variable for Development Endpoint Configuration

Modified the getNehServerEndpoint function to use the NEH_SERVER_ENDPOINT_DEVELOPMENT environment variable for constructing the endpoint URL in development environments. This change ensures that if WORKING_ON_LOCALHOST is set, the application will use the specified development endpoint, and will panic if the variable is not set, aiding in early detection of configuration issues.